+++ example/d2048/game.dart
@@ -9,6 +9,15 @@ part 'model/grid.dart';
 part 'view/board.dart';
 
 main() {
+  /*
+  var tileGrid;
+  String tiles = window.localStorage[Board.jsonName];
+  if (tiles != null) {
+    tileGrid = new TileGrid.fromJsonString(4, tiles);
+  } else {
+    tileGrid = new TileGrid(4);
+  }
+   */
   new Board(querySelector('#canvas'), new TileGrid(4)).draw();
 }